move remaining utils
