+++ packages/renderer/src/config/chains.ts
@@ -103,8 +103,8 @@ export const ChainList = new Map<ChainID, Chain>([
       },
       endpoints: {
         rpcs: [
-          'wss://westend-rpc.polkadot.io',
           'wss://rpc-westend.luckyfriday.io',
+          'wss://westend-rpc.polkadot.io',
           'wss://westend-rpc.dwellir.com',
           'wss://westend-rpc-tn.dwellir.com',
           'wss://rpc.ibp.network/westend',
