+++ packages/renderer/src/renderer/screens/Import/Ledger/Address.tsx
@@ -16,7 +16,7 @@ import { useAddresses } from '@app/contexts/import/Addresses';
 import { useConnections } from '@app/contexts/common/Connections';
 import { useOverlay } from '@polkadot-live/ui/contexts';
 import { chainIcon } from '@ren/config/chains';
-import { getAddressChainId } from '@app/Utils';
+import { getAddressChainId } from '@ren/utils/AccountUtils';
 import type { LedgerAddressProps } from '../types';
 
 export const Address = ({ localAddress, setSection }: LedgerAddressProps) => {
