+++ packages/renderer/src/utils/RenderingUtils.tsx
@@ -1,14 +1,76 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import * as UI from '@polkadot-live/ui/components';
+import PolkadotIcon from '@app/svg/polkadotIcon.svg?react';
+import WestendIcon from '@app/svg/westendIcon.svg?react';
+import KusamaIcon from '@app/svg/kusamaIcon.svg?react';
+
 import BigNumber from 'bignumber.js';
 import { planckToUnit, rmCommas } from '@w3ux/utils';
 import { chainCurrency, chainUnits } from '@ren/config/chains';
+import { useHelp } from '@app/contexts/common/Help';
+import type { AnyData } from '@polkadot-live/types/misc';
 import type { ChainID } from '@polkadot-live/types/chains';
 import type {
   SubscriptionTask,
   TaskCategory,
 } from '@polkadot-live/types/subscriptions';
+
+/**
+ * @name LinksFooter
+ * @summary Render shared footer in views.
+ */
+export const LinksFooter = () => {
+  const { openHelp } = useHelp();
+  return (
+    <UI.LinksFooter
+      handleDisclaimerClick={() => openHelp('help:docs:disclaimer')}
+      handlePrivacyClick={() => openHelp('help:docs:privacy')}
+    />
+  );
+};
+
+/**
+ * @name TooltipWrapper
+ * @summary Utility to render a tooltip over some generic JSX.
+ */
+export const TooltipWrapper = ({
+  theme,
+  wrap,
+  tooltipText = 'Currently Offline',
+  children,
+}: {
+  theme: AnyData;
+  wrap: boolean;
+  tooltipText: string;
+  children: React.ReactNode;
+}) => {
+  if (wrap) {
+    return (
+      <UI.TooltipRx theme={theme} text={tooltipText}>
+        {children}
+      </UI.TooltipRx>
+    );
+  } else {
+    return children;
+  }
+};
+
+/**
+ * @name getIcon
+ * @summary Return the correct network icon based on chain ID.
+ */
+export const getIcon = (chainId: ChainID, iconClass: string) => {
+  switch (chainId) {
+    case 'Polkadot':
+      return <PolkadotIcon className={iconClass} />;
+    case 'Westend':
+      return <WestendIcon className={iconClass} />;
+    case 'Kusama':
+      return <KusamaIcon className={iconClass} />;
+  }
+};
 
 /**
  * @name getTooltipClassForGroup
